# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ilevel API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/users/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Login user",
                "description": "Login user",
                "operationId": "App\\Http\\Controllers\\API\\Auth\\LoginController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/users/logout": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Logout user",
                "description": "Logout user",
                "operationId": "App\\Http\\Controllers\\API\\Auth\\LogoutController::logout",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Bearer lbsiiejroiehfuwoefnelfiower2432420agfasdfsadfasdgrewgwergwergwergewrg",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/users/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Register a new user",
                "description": "Register a new user",
                "operationId": "App\\Http\\Controllers\\API\\Auth\\RegisterController::register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginUserRequest": {
                "title": "Login user request",
                "description": "Login Userrequest body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "email": {
                        "title": "Email",
                        "description": "User email",
                        "type": "string",
                        "example": "John@example.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "User password",
                        "type": "string",
                        "example": "mypassword1"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "User name",
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "title": "Email",
                        "description": "User email",
                        "type": "string",
                        "example": "John@example.com"
                    },
                    "token_type": {
                        "title": "token type",
                        "description": "token type",
                        "type": "string",
                        "example": "Bearer"
                    },
                    "token": {
                        "title": "token",
                        "description": "token",
                        "type": "string",
                        "example": "d;sf;sdbg;afaisfghlierjwoierh7492342pkrof34f845y034u34"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
            "RegisterUserRequest": {
                "title": "Register user request",
                "description": "Register user request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "User name",
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "title": "Email",
                        "description": "User email",
                        "type": "string",
                        "example": "John@example.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "User password",
                        "type": "string",
                        "example": "mypassword1"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "User password",
                        "type": "string",
                        "example": "mypassword1"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}